==> FluxProcessingBB1/1.5.water_table_data_correction.R <==
library(tidyverse)
library(forecast)
library(lubridate)
library(plotly)

rm(list=ls())

# The WaterTable_YK.csv dataset was generated by Johannes
# 
# <Meta info>
# Attached is the water table data. It's a bit more complicated than usual, 
# as the ground level moves in peatland. Hence, I have to run some code before 
# the data is good to be used. The file contains the ground water table elevation 
# above an arbitrary datum. It also contains the ground level above the same datum. 
# "Cground" is the ground level at the center well, where the ground water table elevation 
# is measured automatically by the sensor. The other ones are from the other well+piezometer 
# nests I take my measurements from. All ground elevations are highly correlated. To get ground 
# water table depth you need to subtract "Cground" from "WTheight" . 
# Ground level is not available for the entire time, because I measure it manually and I wasn't 
# in Canada in 2016. You will also notice a difference in data quality. Noisy data is from when 
# we had an old sensor only. A new one was installed in summer 2017. 

## The WTH.csv file generated from WaterTable_YK.csv following Johannes's guide line (subtract "Cground" from "WTheight")

#wth <- read_csv("./Flux-tower/met_data/met_ancillary/wtd/WTH.csv") Need to change path here to run locally
wth <- read_csv("/Users/marionnyberg/Google\ Drive/Micromet\ Lab/Projects/2014-Burns\ Bog/Flux-tower/met_data/met_ancillary/wtd/WTH.csv")

wth$Date <- as.POSIXct(wth$Date,format="%Y-%m-%d %H:%M", tz = "UTC")

# noisy removing
wth$ma_WTH <- ifelse(wth$Date < as.POSIXct("2017-11-01",format="%Y-%m-%d"), ma(wth$WTH,48, centre = TRUE), wth$WTH)

plot_ly(data = wth, x = ~Date, y = ~WTH, name = 'Original data', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~ma_WTH, name = 'Moving average', type = 'scatter', mode = 'lines')

#write_csv(wth,"./Flux-tower/met_data/met_ancillary/wtd/WTH_noisy_removed.csv") Need to change path here to run locally
write_csv(wth, "/Users/marionnyberg/Google\ Drive/Micromet\ Lab/Projects/2014-Burns\ Bog/Flux-tower/met_data/met_ancillary/wtd/WTH_noisy_removed.csv")
