==> FluxProcessingBB1/2.met_filtering_gapfilling.Rmd <==
---
title: "2.met_filtering_gapfilling.R"
author: "Sara Knox"
date: "09/09/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())

library(lubridate)
library(tidyverse)
library(zoo)
library(tseries)
library(plotly)
library(forecast)


# set wd
getwd()
#dir <- "../Flux-tower (1)/"
dir <- "G:/.shortcut-targets-by-id/1txCh9lZ7VGCujXGvBaJCMVnuxT-65q4K/Micromet Lab/Projects/2014-BB1 Burns Bog/Flux-tower (1)"
knitr::opts_knit$set(root.dir = dir)
```

```{r load met_merged.csv, echo=FALSE, include=FALSE}
ts<-read.csv( "./met_data/met_merged/met_merged.csv",sep=",",header=TRUE,dec=".")
head(ts)
tail(ts)
```

```{r gap-fill rows by generating new column, include = FALSE}
# before starting, check the length of df... shouldn't have NA at the end
# also check format of timestamp

ts$Timestamp <- as.POSIXct(ts$Timestamp,format="%Y-%m-%d %H:%M", tz = "UTC")

#to estimate necessary parameters to generate the new empty dataframe with complete rows 
begining<-as.numeric(ts$Timestamp[1])                           #to find the number that is represented by the first data of our time series
as.POSIXct(begining,origin="1970-01-01 00:00:00",tz="UTC") #to confirm that the begining is the right one
ts$Timestamp[1]
     
## Adding year and julian day columns
ts$year <- year(ts$Timestamp)
ts$jday <- yday(ts$Timestamp)
ts$hour_dec <- hour(ts$Timestamp) + minute(ts$Timestamp)/60
 
Ystart<-ts[1,match('year',names(ts))]
Yend<-ts[nrow(ts),match('year',names(ts))]
Dstart<-ts[1,match('jday',names(ts))]
Dend<-ts[nrow(ts),match('jday',names(ts))]

Ystart
Yend
Dstart
Dend

Ndays <- as.numeric((difftime(ts$Timestamp[nrow(ts)],ts$Timestamp[1], "UTC",
         units = c("days"))), units="days")

ts$obs<-c(1:nrow(ts)) 
ts$obs[nrow(ts)]

names(ts)
```

```{r create new dataframe with continuous time step, echo = FALSE}
#To generate a serie of all minutes in a day:
Tsteps <-begining+seq(from=0,to=((Ndays)*(60*60*24)),by=(30*60)) # half an hours data from the begining date to Ndays +1 (that way I assure all measured data are included in the new file) -> I did not use this method
DATE <-as.POSIXct(Tsteps,origin="1970-01-01 00:00:00",tz="UTC")

#Confirming length of the new Time string 
#(THE NEW FILE SHOULD END SAME TIME THE TIME SERIES START AND ONE DAY AFTER THE MEASUREMENTS TIME SERIE)
#DATE[1]
#DATE[length(Tsteps)]
#ts$Timestamp[1]
#ts$Timestamp[length(ts$Timestamp)]
if(DATE[1] ==  ts$Timestamp[1]){c('start dates match')} else {c('check start dates!')}
if(DATE[length(Tsteps)] ==  ts$Timestamp[length(ts$Timestamp)]){c('end dates match')} else {c('check end dates!')}  

#GENERATING A NEW DATA FRAME WITH CONTINUOUS TIME STEPS and data from THE ORIGINAL ONE
cont.DS<-as.data.frame(DATE)
cont.DS[,c(names(ts)[match( "AIR_TEMP_2M",names(ts)):length(names(ts))])]<-NA   #we will just copy the data from the variable "Batt_volt_Min" in ts (variable 3)
cont.DS$DATE<-DATE
#head(cont.DS)

#FILLING THE NEW DATAFRAME WITH DATA FROM THE ORIGINAL DATA FRAME 
for(i in 2:ncol(cont.DS)){  
cont.DS[,i]<-ts[pmatch(cont.DS$DATE,ts$Timestamp),i]           #pmatch look for the observation rows when time columns of both (old and new) dataframes match
                          } 
cont.DS$year <- year(cont.DS$DATE)
cont.DS$jday <- yday(cont.DS$DATE)
cont.DS$hour_dec <- hour(cont.DS$DATE) + minute(cont.DS$DATE)/60
```

```{r apply corrections to met data, echo = FALSE}
# add new corrections here if you notice errors/outliers in met data
DATA<-cont.DS
DATA$obs<-c(1:nrow(cont.DS))
#head(DATA)
#tail(DATA)
DATA <- DATA[-nrow(DATA),]


#######-------------- soil heat flux correction ---------------########


## Soil heat flux correction (for Burns Bog 2017-2018 particular case)
# Did not use SHFP_1 for 2017.09.15-2018.03.07 due to bad quality data
#plot original
plot_ly(data = DATA, x = ~DATE, y = ~SHFP_1, name = 'SHFP 1', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SHFP_2, name = 'SHFP 2', mode = 'lines') %>%
  add_trace(y = ~SHFP_3, name = 'SHFP 3', mode = 'lines')

# Remove bad data 
bad_q <- which(DATA$DATE < "2014-07-10")
DATA$SHFP_1[bad_q] <- NA
DATA$SHFP_2[bad_q] <- NA
DATA$SHFP_3[bad_q] <- NA

bad_q <- which(DATA$DATE >= "2017-09-15" & DATA$DATE < "2018-03-08")
DATA$SHFP_1[bad_q] <- NA

# From 2018.03.07, SHFP_1 should be corrected by muliplying -1
correction <- which(DATA$DATE >= "2018-03-08 14:00:00" & DATA$DATE < "2018-09-07 09:30:00")
DATA$SHFP_1[correction] <- DATA$SHFP_1[correction] * -1

#plot after correction
plot_ly(data = DATA, x = ~DATE, y = ~SHFP_1, name = 'SHFP 1', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SHFP_2, name = 'SHFP 2', mode = 'lines') %>%
  add_trace(y = ~SHFP_3, name = 'SHFP 3', mode = 'lines')


######------------- soil temperature correction ---------########

#plot original

plot_ly(data = DATA, x = ~DATE, y = ~SOIL_TEMP_5CM, name = 'TS 5cm', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_10CM, name = 'TS 10cm', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_50CM, name = 'TS 50cm', mode = 'lines')


bad_q <- c(which(DATA$DATE < "2014-07-10"),
           which(DATA$DATE >= "2014-10-25" & DATA$DATE < "2014-10-27"),
           which(DATA$DATE >= "2014-10-20 18:00:00" & DATA$DATE <= "2014-10-20 22:00:00"),
           which(DATA$DATE >= "2020-01-12 00:00" & DATA$DATE <= "2020-01-16 11:30"), # added by T.S. 2022-01-17
           which(DATA$DATE >= "2020-12-25 00:00" & DATA$DATE <= "2021-01-13 17:30"), # added by T.S. 2022-01-17
           which(DATA$DATE >= "2021-09-21 00:00" & DATA$DATE <= "2021-09-24 11:00"), # added by T.S. 2022-01-17
           which(DATA$DATE >= "2021-10-15 20:00" & DATA$DATE <= "2021-10-19 16:30"), # added by T.S. 2022-01-17
           which(DATA$DATE >= "2021-11-19 10:00" & DATA$DATE <= "2021-11-20 14:00"), # added by T.S. 2022-01-17
           which(DATA$DATE >= "2021-12-07 20:00" & DATA$DATE <= "2021-12-11 00:00")) # added by T.S. 2022-01-17
DATA$SOIL_TEMP_5CM[bad_q] <- NA
DATA$SOIL_TEMP_10CM[bad_q] <- NA
DATA$SOIL_TEMP_50CM[bad_q] <- NA

# convert to timeseries class
ts.TS_5<- ts(DATA$SOIL_TEMP_5CM) 
ts.TS_10<- ts(DATA$SOIL_TEMP_10CM) 
ts.TS_50<- ts(DATA$SOIL_TEMP_50CM) 

# # remove spikes
# ts.TS_5.outlier <- tsoutliers(ts.TS_5, iterate = 4) %>% as.data.frame()
# DATA$SOIL_TEMP_5CM[ts.TS_5.outlier$index] <- ts.TS_5.outlier$replacements
# 
# ts.TS_10.outlier <- tsoutliers(ts.TS_10, iterate = 4) %>% as.data.frame()
# DATA$SOIL_TEMP_10CM[ts.TS_10.outlier$index] <- ts.TS_10.outlier$replacements
# 
# ts.TS_50.outlier <- tsoutliers(ts.TS_50, iterate = 4) %>% as.data.frame()
# DATA$SOIL_TEMP_50CM[ts.TS_50.outlier$index] <- ts.TS_50.outlier$replacements

# fill short gaps with linear interpolation
DATA$SOIL_TEMP_5CM[which(DATA$DATE > "2015-07-29")] <- as.numeric(tsclean(ts.TS_5))[which(DATA$DATE > "2015-07-29")]
DATA$SOIL_TEMP_10CM[which(DATA$DATE > "2015-07-29")] <- as.numeric(tsclean(ts.TS_10))[which(DATA$DATE > "2015-07-29")]
DATA$SOIL_TEMP_50CM[which(DATA$DATE > "2016-01-16")] <- as.numeric(tsclean(ts.TS_50))[which(DATA$DATE > "2016-01-16")]

#plot after correction
plot_ly(data = DATA, x = ~DATE, y = ~SOIL_TEMP_5CM, name = 'TS 5cm', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_10CM, name = 'TS 10cm', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_50CM, name = 'TS 50cm', mode = 'lines')



######------------ Radiation corrections ---------------------########


## Corrections for Radiation measurements - SEE IF CORRECTIONS ARE ALSO NEEDED FOR LW_IN, SW_IN
#plot original
plot_ly(data = DATA, x = ~DATE, y = ~SHORTWAVE_OUT, name = 'SW_OUT', type = 'scatter', mode = 'lines') 

bad_q <- which(DATA$DATE < "2014-07-10")
DATA$SHORTWAVE_OUT[bad_q] <- NA

#plot after correction
plot_ly(data = DATA, x = ~DATE, y = ~SHORTWAVE_OUT, name = 'SW_OUT', type = 'scatter', mode = 'lines') 

## Wind speed correction (for Burns Bog 2017-2018 particular case)
# Did not use cup wind speed for 2017.07.04-2018.02.09 due to bad quality data
bad_q <- which(DATA$DATE >= "2017-07-04" & DATA$DATE < "2018-02-10")
DATA$WIND_VELOCITY_CUP[bad_q] <- NA

# Replace this gap using EC wind speed and then remove EC wind speed column
DATA$WIND_VELOCITY_CUP[bad_q] <- DATA$WIND_VELOCITY_EC2[bad_q]
DATA <- DATA[, -grep("WIND_VELOCITY_EC2", colnames(DATA))]

plot_ly(data = DATA, x = ~DATE, y = ~WIND_VELOCITY_CUP, type = 'scatter', mode = 'lines')



########------------------ WTH correction -------------------########

# read noisy removed wth 
# wth <- read_csv("G:/.shortcut-targets-by-id/1txCh9lZ7VGCujXGvBaJCMVnuxT-65q4K/Micromet Lab/Projects/2014-Burns\ Bog/Flux-tower (1)/met_data/met_ancillary/wtd/WTH_noisy_removed.csv")
# we don't use this one because it is WTH relative to moving bog height

# use moving average to remove noise
DATA$WTH <- ifelse(DATA$WTH < as.POSIXct("2017-11-01",format="%Y-%m-%d"), ma(DATA$WTH,48, centre = TRUE), DATA$WTH)

plot_ly(data = DATA, x = ~DATE, y = ~WTH, name = 'noisy removed', type = 'scatter', mode = 'lines')

# Create new column for WTH
DATA$WTH_man <- NA  
DATA$WTH_absolute <- NA

# linear regression for different periods from JE
WTH_corr_coeff <- data.frame(slope = c(0.0104, 0.0035, 0.0094), 
                             int = c(3.3028, 2.6401, 3.6561), 
                             period = c("before2017", "mid2017", "after2017"))

# assign period
before2017 <- which(DATA$DATE >= "2015-07-01 00:00:00" & DATA$DATE < "2017-01-13 02:30:00")
mid2017 <-  which(DATA$DATE >= "2017-01-13 02:30:00" & DATA$DATE <= "2017-10-26 18:30:00")
after2017 <- which(DATA$DATE > "2017-10-26 18:30:00")

# calculate WTH_man = WTH_PT * slope + int (result is in m a.s.l)
DATA$WTH_man[before2017] <- (DATA$WTH[before2017]) * WTH_corr_coeff[1,1] + WTH_corr_coeff[1,2]
DATA$WTH_man[mid2017] <- (DATA$WTH[mid2017] * WTH_corr_coeff[2,1]) + WTH_corr_coeff[2,2]
DATA$WTH_man[after2017] <- (DATA$WTH[after2017] * WTH_corr_coeff[3,1]) + WTH_corr_coeff[3,2]

# plot WTH (in m a.s.l) 
plot_ly(data = DATA, x = ~DATE, y = ~WTH_man, name = 'WTD', type = 'scatter', mode = 'lines') %>% 
  layout(xaxis = list(title = "Date"), yaxis = list(title = "WTD (m.a.s.l)"))
  
# plot WTH relative to different bog surfaces
plot_ly(data = DATA, x = ~DATE, y = ~WTH_man - 3.36 , name = 'WTD using bog height median = 3.36 masl', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~WTH_man - 3.452 , name = 'WTD using bog height GNSS survey = 3.452 masl', mode = 'lines') %>% 
  add_trace(y = ~WTH_man - (4.473 - 1.18) , name = 'WTD using bog height from MN = 3.293 masl', mode = 'lines') %>%
  layout(xaxis = list(title = "Date"), yaxis = list(title = "WTD absolute (m)"))

# adding absolute WTH to the table (WTH in m)
DATA$WTH_absolute <- DATA$WTH_man - 3.293


# Cleaning WTD 

# WTD in 2015 looks much lower. Since it's from the same sensor, the slope is probably the same with WTD from 2016. 
# I assume that we just have to correct for the offset 

y2015 <- which(DATA$DATE <= "2015-12-11 00:00:00")
DATA$WTH_absolute[y2015] <- DATA$WTH_absolute[y2015] + 0.1624634

# # Remove spikes in Dec 11 2015 & 2016 Aug 29 2016
# DATA$WTH_absolute[which(DATA$DATE >= as.POSIXct("2015-12-11 00:00:00", tz = "UTC") & DATA$DATE < as.POSIXct("2015-12-12 00:00:00", tz = "UTC"))] <- NA 
# DATA$WTH_absolute[which(DATA$DATE >= as.POSIXct("2016-08-28 11:30:00", tz = "UTC") & DATA$DATE <  as.POSIXct("2016-08-30 00:00:00", tz = "UTC"))] <- NA 

# remove spikes
ts.wth <- ts(DATA$WTH_absolute) 
ts.wth.outlier <- tsoutliers(ts.wth, iterate = 4) %>% as.data.frame()
DATA$WTH_absolute[ts.wth.outlier$index] <- ts.wth.outlier$replacements

# fill short gaps with linear interpolation
DATA$WTH_absolute[which(DATA$DATE > "2015-07-29")] <- as.numeric(tsclean(ts.wth))[which(DATA$DATE > "2015-07-29")]

plot_ly(data = DATA, x = ~DATE, y = ~WTH_absolute, name = 'WTD', type = 'scatter', mode = 'lines') %>% 
  layout(xaxis = list(title = "Date"), yaxis = list(title = "WTD (m)"))



########------------------ PA correction ------------------########

# Combined PA measurements
#PA_EC1 <- which(DATA$DATE <= "2016-06-16") PA_EC1 - not in csv files from script 0 and 1
#DATA$PA_2M[PA_EC1] <- DATA$PA_EC1[PA_EC1]

plot_ly(data = DATA, x = ~DATE, y = ~PA_2M, name = 'Original', type = 'scatter', mode = 'lines')



########------------------ PAR correction ------------------########

# Correction for PAR data from 2019 (not sure why I had to split it over 2 lines, but the first is back converting to mV and then applying the new calibration)
correction <- which(DATA$DATE >= "2019-01-01 00:30:00")
DATA$INCOMING_PAR[correction] <- (DATA$INCOMING_PAR[correction]/287.59)
DATA$INCOMING_PAR[correction] <- (DATA$INCOMING_PAR[correction] * 462.87)

## 2nd correction made because 1st one still looked too high
correction2 <- which(DATA$DATE < "2019-01-01")
DATA$INCOMING_PAR[correction2] <- DATA$INCOMING_PAR[correction2]/287.59
DATA$INCOMING_PAR[correction2] <- DATA$INCOMING_PAR[correction2] * 375

plot_ly(data = DATA, x = ~DATE, y = ~INCOMING_PAR, name = 'PAR with correction', type = 'scatter', mode = 'lines')



######------------------ SVWC correction ------------------#########
# added by T.S. 2022-01-17

#plot original

plot_ly(data = DATA, x = ~DATE, y = ~SVWC, name = 'SVWC', type = 'scatter', mode = 'lines')

swc <- DATA %>% select(DATE, SVWC)
plot_ly(data = swc, x = ~DATE, y = ~SVWC, name = 'SVWC', type = 'scatter', mode = 'lines')

# remove bad SVWC 
bad_SVWC <- which(DATA$DATE >= "2021-10-13")
DATA$SVWC[bad_SVWC] <- NA

# SVWC periods based on the sensor position (see Burns Bog field site log)
y2016_1 <- which(swc$DATE >= '2014-12-13' & swc$DATE <= '2016-09-16')
y2016_2 <- which(swc$DATE >= '2016-11-04' & swc$DATE <= '2017-02-03')
y2017_1 <- which(swc$DATE >= '2017-03-21 11:30' & swc$DATE <= '2017-07-13 10:30')
y2017_2 <- which(swc$DATE >= '2017-07-13 14:30' & swc$DATE <= '2017-10-12')
y2017_3 <- which(swc$DATE >= '2017-11-03' & swc$DATE <= '2017-11-21')
y2018_1 <- which(swc$DATE >= '2018-02-09' & swc$DATE <= '2018-06-19')
y2018_2 <- which(swc$DATE >= '2018-07-05' & swc$DATE <= '2018-10-02')
y2018_3 <- which(swc$DATE >= '2018-10-04' & swc$DATE <= '2018-11-13')
y2019_1 <- which(swc$DATE >= '2019-05-23 12:30' & swc$DATE <= '2020-11-12')
y2021_1 <- which(swc$DATE >= '2021-06-14' & swc$DATE < '2021-11-20')
y2022 <- which(swc$DATE >= "2022-02-01")


# adjust offset (there's no way to know the exact numbers)
swc$vol.SM <- NA
swc$vol.SM[y2016_1] <- (swc$SVWC[y2016_1]) * 2.7 - 159
swc$vol.SM[y2016_2] <- swc$SVWC[y2016_2]
swc$vol.SM[y2017_1] <- swc$SVWC[y2017_1]
swc$vol.SM[y2017_2] <- swc$SVWC[y2017_2] - 17
swc$vol.SM[y2017_3] <- swc$SVWC[y2017_3] - 7
swc$vol.SM[y2018_1] <- swc$SVWC[y2018_1] * 0.8 + 23.8
swc$vol.SM[y2018_2] <- swc$SVWC[y2018_2] - 11.5
swc$vol.SM[c(y2018_3, y2019_1, y2021_1, y2022)] <- swc$SVWC[c(y2018_3, y2019_1, y2021_1, y2022)]


plot_ly(data = swc, x = ~DATE, y = ~vol.SM, name = 'SVWC', type = 'scatter', mode = 'lines')

library(fpp2)


# convert to timeseries class
ts.swc <- ts(swc$vol.SM) 

# plot outliers to check
autoplot(tsclean(ts.swc), series="clean", color='red', lwd=0.9) +
  autolayer(ts.swc, series="original", color='gray', lwd=1) +
  geom_point(data = tsoutliers(ts.swc) %>% as.data.frame(), 
             aes(x=index, y=replacements), col='blue') +
  labs(x = "Day", y = "SWC")

# replace spikes
ts.swc.outlier <- tsoutliers(ts.swc, iterate = 4) %>% as.data.frame()
ts.swc.outlier <- ts.swc.outlier[1:638, ]  # refine because the data looks good, should edit this once SWC sensor is replaced
swc$vol.SM.fill <- swc$vol.SM
swc$vol.SM.fill[ts.swc.outlier$index] <- ts.swc.outlier$replacements

DATA$SVWC <- swc$vol.SM.fill

# fill long gaps with SWC at saturation as the mean of max. 1% soil moisture on record 
theta.sat <- as.numeric(quantile(swc$vol.SM.fill, .99, na.rm = T))
m <- is.na(DATA$SVWC) == T & DATA$WTH_absolute > 0 & DATA$DATE >  "2014-12-13"
DATA$SVWC <- ifelse(m , theta.sat, DATA$SVWC)

# fill short gaps with linear interpolation
DATA$SVWC[which(DATA$DATE > "2015-07-29")] <- as.numeric(tsclean(ts(DATA$SVWC)))[which(DATA$DATE > "2015-07-29")]


plot_ly(data = DATA, x = ~DATE, y = ~SVWC, name = 'SVWC', type = 'scatter', mode = 'lines')




```

```{r gap-fill using ECCC data & ClimateAtUBC data, echo = FALSE}
#Gapfilling using environment canada dataset (Burns Bog site) & climateATUBC (http://ibis.geog.ubc.ca/~achristn/data.html; http://137.82.254.122/DataExport.aspx)

no_gapfilled <- DATA

##### 1. environment canada dataset: airT, RH, wind speed #####
near_station <- read_csv(paste0(dir, "/met_data/met_ancillary/ECCC/ECCC_met_merged.csv"))
#names(near_station)

for (i in c(1, 2, 4, 5)){
	var <- names(near_station)[i+5]
	gap_row <- which(is.na(DATA[, grep(var, colnames(DATA))]))
	intersect_row_for_near_station <- which(near_station$Timestamp %in% DATA$DATE[gap_row])
	intersect_row_for_DATA <- which(DATA$DATE %in% near_station$Timestamp[intersect_row_for_near_station])
	DATA[intersect_row_for_DATA, grep(var, colnames(DATA))] <- near_station[intersect_row_for_near_station, i+5]
	# linear interporation: environment canada data is 1 hour, so need to interpolated to make 30 min dataset
	DATA[, grep(var, colnames(DATA))] <- na.approx(DATA[, grep(var, colnames(DATA))], na.rm = FALSE, rule = 2)
}

#plotting
plot_ly(data = DATA, x = ~DATE, y = ~AIR_TEMP_2M, name = 'gap-filled', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~AIR_TEMP_2M, name = 'non gap-filled', type = 'scatter', mode = 'lines') 

plot_ly(data = DATA, x = ~DATE, y = ~RH_2M, name = 'gap-filled', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~RH_2M, name = 'non gap-filled', type = 'scatter', mode = 'lines') 

plot_ly(data = DATA, x = ~DATE, y = ~WIND_VELOCITY_CUP, name = 'gap-filled', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~WIND_VELOCITY_CUP, name = 'non gap-filled', type = 'scatter', mode = 'lines') 

plot_ly(data = DATA, x = ~DATE, y = ~PA_2M, name = 'gap-filled', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~PA_2M, name = 'non gap-filled', type = 'scatter', mode = 'lines') 

##### 2. climateATUBC dataset: Rg, ppt ####
#(http://137.82.254.122/DataExport.aspx) from Jan 1, 2014 to 2021 & change extension to csv
#ubc <- read_csv("./met_data/met_ancillary/ubc_climate/climateAtUBC.csv")
ubc <- read_csv(paste0(dir, "/met_data/met_ancillary/ubc_climate/climateAtUBC.csv"))

# Totem field data is no longer available on the web since mid 2021. Download from this link instead: https://link.landfood.ubc.ca/biomet/users/TotemDataForTin.dat 
# delete the first row of the dataset manually first before reading here
ubc2 <- read.table(paste0(dir, "/met_data/met_ancillary/ubc_climate/TotemDataForTin.txt"), header = T, sep = ",")

colnames(ubc2) <- colnames(ubc)

# Date character -> Date data
ubc$DATE <- as.POSIXct(ubc$time_stamp,format="%m/%d/%Y %I:%M:%S %p", tz = "UTC")
ubc2$DATE <- as.POSIXct(ubc2$time_stamp, format ="%Y-%m-%d %H:%M:%S", tz = "UTC" )

ubc <- rbind(ubc, ubc2)

##

gap <- which(is.na(ubc$DATE))
ubc$DATE[gap] <- as.POSIXct(ubc$time_stamp,format="%m/%d/%Y", tz = "UTC")[gap]

#subset
st <- which(ubc$DATE== DATA$DATE[1])
la <- which(ubc$DATE == DATA$DATE[nrow(DATA)])
ubc_use <- ubc[c(st:la),]

ubc_use$global_radiation_main[ubc_use$global_radiation_main == "-999"]<-NA

# Because this data tower is pretty far from the BB tower, check relationships
#### 2.1. Rg #####
#linear regression
a<-round(coef(lm(DATA$SHORTWAVE_IN~ubc_use$global_radiation_main))[2],digits=2) #coef(lm(y~x))[2] is the slope of the regression
b<-round(coef(lm(DATA$SHORTWAVE_IN~ubc_use$global_radiation_main))[1],digits=2)   
r2<-round(summary(lm(DATA$SHORTWAVE_IN~ubc_use$global_radiation_main))$ r.squared,digits=2)
lm_eq<-paste0("y=",a,"x",ifelse(b>0,"+",""),b)
R2<-bquote(R^2 == .(r2)) 

#gapfilling
DATA$SHORTWAVE_IN[which(is.na(DATA$SHORTWAVE_IN))] <- ubc_use$global_radiation_main[which(is.na(DATA$SHORTWAVE_IN))]*a + b

# Linearly interpolate small gaps where ubc data was -999
DATA$SHORTWAVE_IN <- na.approx(DATA$SHORTWAVE_IN, na.rm = FALSE, rule = 2)

#Plot:
plot(ubc_use$global_radiation_main,no_gapfilled$SHORTWAVE_IN, main = "Global radiation (Short wave in)", xlab = "UBC climate site", ylab = "BB flux tower")
abline(0,1,col='grey',lty=2)
abline(lm(DATA$SHORTWAVE_IN~ubc_use$global_radiation_main),col='red')
mtext( lm_eq,line = -1, side = 3)
mtext(R2,line=-2, side = 3)

#Plot time series:
plot_ly(data = no_gapfilled, x = ~DATE, y = ~SHORTWAVE_IN, name = 'BB tower', type = 'scatter', mode = 'lines') %>%
  add_trace(data = ubc_use, x = ~DATE, y = ~global_radiation_main, name = 'UBC tower', type = 'scatter', mode = 'lines') 

# gapfilling plot
plot_ly(data = DATA, x = ~DATE, y = ~SHORTWAVE_IN, name = 'gap-filled', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~SHORTWAVE_IN, name = 'original', type = 'scatter', mode = 'lines') 

#### 2.2. Precipitation #####
#gapfilling
ubc_use$precipitation_main[ubc_use$precipitation_main == "-999"]<-NA

DATA$PRECIP[which(is.na(DATA$PRECIP))] <- ubc_use$precipitation_main[which(is.na(DATA$PRECIP))]

# Linearly interpolate small gaps where ubc data was -999
DATA$PRECIP <- na.approx(DATA$PRECIP, na.rm = FALSE, rule = 2)

#plot:
plot_ly(data = ubc_use, x = ~DATE, y = ~precipitation_main, name = 'BB', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~PRECIP, name = 'UBC', type = 'scatter', mode = 'lines') 

plot_ly(data = DATA, x = ~DATE, y = ~PRECIP, name = 'gap-filled', type = 'scatter', mode = 'lines') %>%
  add_trace(data = no_gapfilled, x = ~DATE, y = ~PRECIP, name = 'original', type = 'scatter', mode = 'lines') 


```

```{r create new output variables, echo = FALSE}
### Make new column for net radiation, soil heat flux, available energy, es, vpd

DATA$NR <- (DATA$SHORTWAVE_IN - DATA$SHORTWAVE_OUT) + 
  (DATA$LONGWAVE_IN - DATA$LONGWAVE_OUT)

DATA$G<-rowMeans(DATA[,grep('SHFP',names(DATA))],na.rm = TRUE)

DATA$AE<-DATA$NR-DATA$G

DATA$es <- 0.611*exp(17.502*DATA$AIR_TEMP_2M/(DATA$AIR_TEMP_2M+240.97))
DATA$VPD <- DATA$es*(1-(DATA$RH_2M/100))

### Calculate SG (heat storage flux in the soil above the soil heat fluxes measurement - W m^-2) -- TS (2021-10-05)

dT_dt <- diff(DATA$SOIL_TEMP_5CM) / 1800
dZ <- 0.05
SVWC.99 <- quantile(DATA$SVWC/100, 0.99, na.rm = T)
Vo <- 1 - SVWC.99
C_soil <- DATA$SVWC/100 * 4.18 + 3.14 * Vo + 0.0012 * (1 - DATA$SVWC/100 - Vo) 
SG <- C_soil * 10e6 * dT_dt * dZ 

# add heat storage column
DATA$SG <- SG 

# remove values that are off due to Ts error
DATA$SG <- ifelse(DATA$SG >= quantile(DATA$SG, .97, na.rm = T) | 
                  DATA$SG <= quantile(DATA$SG, .25, na.rm = T), 
                  NA, DATA$SG) 

# calculate corrected G
DATA$G_corr <- ifelse(is.na(DATA$SG), DATA$G, DATA$G + DATA$SG)

plot_ly(data = DATA, x =~DATE,  y= ~G_corr, name = 'G_corr', type = 'scatter', mode = 'lines') %>% 
  add_trace(x = ~DATE, y = ~G,name = 'G', type = 'scatter', mode = 'lines')
  

plot_ly(data = DATA, x = ~DATE, y = ~DATA$NR, name = 'NETRAD', type = 'scatter', mode = 'lines') %>%
  add_trace(x = ~DATE, y = ~G,name = 'G', type = 'scatter', mode = 'lines') %>%
  add_trace(x = ~DATE, y = ~AE, name = 'AE', type = 'scatter', mode = 'lines') 

# Plot all variables
for (i in 2:ncol(DATA)){
	use <- DATA[,i]
	
	plot_ly(data = DATA, x = ~DATE, y = ~use, type = 'scatter', mode = 'lines') %>%
	  layout(yaxis = list(title = names(DATA)[i]))
}
```

```{r save filtered & gap-filled output, echo = FALSE}
# save output
write.csv(DATA,paste0(dir, '/met_data/met_merged/met_corrected_gapfilled.csv'),row.names=FALSE)      
```
